refactor(settings): theme row spans full width with live ColorSwatches
- Extract ColorSwatches + SWATCH_KEYS from CThemeList into a shared
  components/commands/renders/theme/ThemeColorSwatches.tsx so both the
  \`theme list\` command renderer and the settings dialog can reuse the
  same swatch strip
- TerminalSettingsDialog grid now reads:
    Language    | Font
    Scrollback  | Font Size
    Theme select|  ColorSwatches  ← full-width row, selector on left,
                                    live swatches of the active palette
                                    on right
  The swatches reflect the currently-applied theme (from
  useThemeEngine().palette) so changing the dropdown updates them in
  place — quick visual feedback without leaving the dialog

Build, typecheck and lint all clean.
